requirements.txt
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
click==8.0.1
colorama==0.4.4
DateTime==4.3
Flask==2.0.1
Flask-Cors==3.0.10
geos==0.2.3
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
pyclipper==1.3.0
pycparser==2.20
pymongo==3.12.0
pytz==2021.1
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wincertstore==0.2
zope.interface==5.4.0
